requirements.txt
asgiref==3.7.2
Django==5.0.1
django-pwa==1.1.0
django-sslserver==0.22
pillow==10.2.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.9.0
